refactor(web): better handling of files
