+++ web/package.json
@@ -42,7 +42,7 @@
     "test": "react-scripts test",
     "eject": "react-scripts eject",
     "lint:check": "eslint src --ext .ts,.tsx",
-    "lint:": "eslint src --ext .ts,.tsx --fix",
+    "lint": "eslint src --ext .ts,.tsx --fix",
     "cypress:open": "cypress open",
     "cypress:run": "cypress run"
   },
